Creating an object for books
Great job creating a few data types and functions. I added an object that will make it easier to work with the book data. I then created an empty an array called shopping cart so that we can add books to the cart by using the .push command inside the addToCart function.